chore: move otp mailsender interface to otp package
